requirements.txt
opencv-python-headless==4.8.0.74
mediapipe==0.10.7
numpy==1.24.4
absl-py==1.4.0
